requirements.txt
flask
fastapi
moviepy
starlette
pathlib
progressist
pydantic~=1.10.2
pySmartDL
python-dotenv
python-jose[all]
requests
requests-toolbelt
urllib3
uvicorn
